added access logic by username

+++ user-app/src/main/java/org/alexov/otus/user/UserController.java
@@ -2,9 +2,16 @@ package org.alexov.otus.user;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.coyote.BadRequestException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.AuthenticatedPrincipal;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
+
+import java.nio.file.AccessDeniedException;
+import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
@@ -12,10 +19,26 @@ public class UserController {
 
     private final UserRepository userRepository;
     private final CredRepo credRepo;
+    private final PasswordEncoder passwordEncoder;
 
     @GetMapping("/user/{id}")
-    public ResponseEntity<?> findUserById(@PathVariable Long id) {
-        return ResponseEntity.ok(userRepository.findById(id));
+    public ResponseEntity<?> findUserById(@PathVariable Long id, Authentication principal) throws AccessDeniedException {
+        String username = principal.getName();
+        Optional<OtusUser> byId = userRepository.findById(id);
+        ResponseEntity<String> FORBIDDEN = isForbidden(byId, username);
+        if (FORBIDDEN != null) return FORBIDDEN;
+
+        return ResponseEntity.ok(byId);
+    }
+
+    private static ResponseEntity<String> isForbidden(Optional<OtusUser> byId, String username) {
+        try{
+            byId.map(OtusUser::getUsername).filter(u -> u.equals(username))
+                    .orElseThrow(() -> new AccessDeniedException("Access denied"));
+        } catch (AccessDeniedException e) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(e.getMessage());
+        }
+        return null;
     }
 
     @PostMapping("/user")
@@ -29,8 +52,11 @@ public class UserController {
 
     @PutMapping("/user/{id}")
     public ResponseEntity<?> updateUser(@PathVariable Long id, @RequestBody OtusUser user, Authentication authentication) throws BadRequestException {
-        var principal = authentication.getPrincipal().toString();
-        OtusUser otusUser = userRepository.findById(id).orElseThrow();
+        var otusUserOptional = userRepository.findById(id);
+        ResponseEntity<String> FORBIDDEN = isForbidden(otusUserOptional, authentication.getName());
+        if (FORBIDDEN != null) return FORBIDDEN;
+
+        var otusUser = otusUserOptional.orElseThrow();
         otusUser.setUsername(user.getUsername());
         otusUser.setFirstName(user.getFirstName());
         otusUser.setLastName(user.getLastName());
@@ -40,7 +66,11 @@ public class UserController {
     }
 
     @DeleteMapping("/user/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<?> deleteUser(@PathVariable Long id, Authentication authentication) {
+        var otusUserOptional = userRepository.findById(id);
+        ResponseEntity<String> FORBIDDEN = isForbidden(otusUserOptional, authentication.getName());
+        if (FORBIDDEN != null) return FORBIDDEN;
+
         userRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
@@ -50,7 +80,15 @@ public class UserController {
         if(credRepo.findById(user.getUsername()).isPresent()){
             throw new BadRequestException("Username is already in use");
         }
-        userRepository.save(OtusUser.builder().username(user.getUsername()).credentials(user).build());
-        return ResponseEntity.ok().build();
+        OtusUser save = userRepository.save(OtusUser.builder().username(user.getUsername()).credentials(UserCred.builder()
+                .username(user.getUsername())
+                .password(passwordEncoder.encode(user.getPassword()))
+                .build()).build());
+        return ResponseEntity.ok(save);
+    }
+
+    @GetMapping("/registration")
+    public ResponseEntity<?> registerUser() {
+        return ResponseEntity.ok("test");
     }
 }
